Extract ComparisonState hook out to its own file to fix fast refresh.
